Update structure #2

+++ app/controllers/BaseController.js
@@ -5,7 +5,7 @@ class BaseController {
     this.model = model;
   }
 
-  async getAll(req, res) {
+  async getAdmin(req, res) {
     try {
       const data = await this.model.findWithDeleted();
       res.status(200).json(data);
@@ -24,9 +24,27 @@ class BaseController {
     }
   }
 
+  async getListByIds(req, res) {
+    if (!req.query.ids) {
+      return res.status(404).json({ message: 'IDS is required' });
+    }
+    const ids = req.query.ids.split(',');
+    try {
+      const options = req.paginateOptions;
+      const query = { _id: { $in: ids } };
+      const data = await this.model.paginate(query, options);
+      res.status(200).json(data);
+    } catch (error) {
+      res.status(500).json(error.message);
+    }
+  }
+
   async getByParam(req, res) {
     try {
       const param = req.params.param;
+      if (!param) {
+        return res.status(404).json({ message: 'Param is required' });
+      }
       let data;
 
       if (mongoose.Types.ObjectId.isValid(param)) {
@@ -56,6 +74,10 @@ class BaseController {
   }
 
   async update(req, res) {
+    const { id } = req.params;
+    if (!id) {
+      return res.status(404).json({ message: 'ID is required' });
+    }
     try {
       const data = await this.model.findByIdAndUpdate(
         req.params.id,
@@ -70,6 +92,10 @@ class BaseController {
   }
 
   async delete(req, res) {
+    const { id } = req.params;
+    if (!id) {
+      return res.status(404).json({ message: 'ID is required' });
+    }
     try {
       await this.model.delete({ _id: req.params.id });
       res.status(200).json({ message: 'Deleted successfully' });
@@ -81,6 +107,10 @@ class BaseController {
   async deleteMany(req, res) {
     try {
       const { ids } = req.body;
+      if (!ids) {
+        return res.status(404).json({ message: 'IDS is required' });
+      }
+
       await this.model.delete({ _id: { $in: ids } });
       res.status(200).json({ message: 'Deleted successfully' });
     } catch (error) {
@@ -100,6 +130,10 @@ class BaseController {
   }
 
   async restore(req, res) {
+    const { id } = req.params;
+    if (!id) {
+      return res.status(404).json({ message: 'ID is required' });
+    }
     try {
       const data = await this.model.restore({ _id: req.params.id });
       res.status(200).json(data);
@@ -109,8 +143,12 @@ class BaseController {
   }
 
   async forceDelete(req, res) {
+    const { id } = req.params;
+    if (!id) {
+      return res.status(404).json({ message: 'ID is required' });
+    }
     try {
-      await this.model.findByIdAndDelete(req.params.id);
+      await this.model.findByIdAndDelete(id);
       res.status(200).json({ message: 'Deleted successfully' });
     } catch (error) {
       res.status(500).json(error.message);
@@ -119,6 +157,9 @@ class BaseController {
 
   async forceDeleteMany(req, res) {
     const { ids } = req.body;
+    if (!ids) {
+      return res.status(404).json({ message: 'IDS is required' });
+    }
     try {
       await this.model.deleteMany({ _id: { $in: ids } });
       res.status(200).json({ message: 'Deleted successfully' });
